chore: Adjust initialization to be available on all Uno targets
