+++ src/Uno.UWP/System/Power/PowerManager.cs
@@ -112,26 +112,6 @@ public static partial class PowerManager
 	}
 #endif
 
-	/// <summary>
-	/// Initializes the PowerManager.
-	/// </summary>
-	/// <returns>A value indicating whether the initialization succeeded.</returns>
-	public static Task<bool> InitializeAsync()
-	{
-#if __WASM__
-		return InitializePlatformAsync();
-#elif __ANDROID__ || __IOS__
-		return Task.FromResult(true);
-#else
-		if (typeof(PowerManager).Log().IsEnabled(LogLevel.Error))
-		{
-			typeof(PowerManager).Log().LogError("PowerManager is not implemented on this platform", ex);
-		}
-
-		return Task.FromResult(false);
-#endif
-	}
-
 	internal static void RaiseRemainingChargePercentChanged()
 	{
 		var currentValue = RemainingChargePercent;
